fix: GitHub sync (PR overflow, redirects) and last-synced UI
- GitHub: use BigInteger for github_pr_id (fixes int32 overflow for large PR IDs)
- GitHub: add follow_redirects to httpx for renamed repos (301 handling)
- GitHub: fix multiple-rows in identity mapping queries
- Migration 095: rename integrations.provider to connector
- Migration 096: change github_pull_requests.github_pr_id to BIGINT
- DataSources: increase sync poll timeout to ~5min, always refetch on completion
- Auth: when multiple integrations exist, show the one with most recent last_sync_at
- Config: add optional MIGRATION_DATABASE_URL for migrations

Made-with: Cursor

+++ backend/models/github_pull_request.py
@@ -9,7 +9,7 @@ import uuid
 from datetime import datetime
 from typing import TYPE_CHECKING, Any, Optional
 
-from sqlalchemy import DateTime, ForeignKey, Index, Integer, String, Text
+from sqlalchemy import BigInteger, DateTime, ForeignKey, Index, Integer, String, Text
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -49,8 +49,8 @@ class GitHubPullRequest(Base):
         UUID(as_uuid=True), ForeignKey("github_repositories.id"), nullable=False
     )
 
-    # PR identification
-    github_pr_id: Mapped[int] = mapped_column(Integer, nullable=False)
+    # PR identification (BIGINT: GitHub IDs can exceed int32)
+    github_pr_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
     number: Mapped[int] = mapped_column(Integer, nullable=False)
 
     # Content
